✨ (models): add event_tags_joined field to Event model
This commit introduces a new field, `event_tags_joined`, to the
`Event` model. This field stores a comma-separated string of all
event tag names associated with an event. The `set_browsable`
callback is updated to set this field after an event is saved. This
change allows for easier searching and filtering of events based on
their tags.