🚚 (tasks.rake): move code_to_markdown task to late:other namespace 🔥 (tasks.rake): remove dump namespace
The 'dump' namespace was removed because it contained a task
'code_to_markdown' that was not actively used and was creating
unnecessary clutter in the rake task list. The
'code_to_markdown' task has been moved to the 'late:other'
namespace to keep it available if needed, but to signal that it is
not part of the main task flow.