🎨 (purchase/_ticket_type.erb): change col class to col-auto
The change was made to improve the layout of the ticket type
selection on the purchase page. By changing the `col` class to
`col-auto`, the width of the quantity selection element will be
determined by its content, rather than taking up the full width of
its parent. This allows the other elements in the row to be
displayed more naturally.