🐛 (models): fix issue with duplicate emails and downcase email
The `event_emails` method in `Pmail` was not mapping the tickets
to emails, resulting in the `emails.include?(ticket.email)` check
always returning false. This led to duplicate emails being sent.
This commit fixes this by mapping the tickets to emails before
checking for inclusion.

The email address was not downcased and stripped before validation
in the Ticket model. This commit fixes this by downcasing and
stripping the email address before validation.