🔥 (app/views/events/_feedback.erb): remove condition that checks if activities are empty
The condition `@event.organisation.activities.empty?` was removed because it was redundant and unnecessary. The primary check is whether there are any unscoped event feedbacks for the organization. Whether or not the organization has activities is irrelevant to displaying the feedback. Removing this condition simplifies the logic and ensures that feedback is displayed as long as it exists, regardless of the presence of activities.