✨ (models/event.rb): consider sales_end when checking availability
This commit modifies the `sold_out?` and `tickets_available?` methods
in the Event model to consider the `sales_end` attribute of ticket
types.

Previously, these methods only checked the
`number_of_tickets_available_in_single_purchase` to determine if an
event was sold out or if tickets were available. Now, they also check
if the `sales_end` date has passed. If `sales_end` is in the past, the
ticket type is considered unavailable, even if there are tickets
remaining.

This change ensures that tickets are no longer sold after the `sales_end`
date, even if the ticket quantity has not been exhausted.