✨ Add customizable theme color for embedded events
Add a `theme_color` field to the Organisation model and form.
Create a new SCSS file (`organisation.scss`) with styles using a
`$theme-color` variable.
Implement a controller action to dynamically compile `organisation.scss`
into CSS, allowing the `$theme-color` to be set via a URL parameter.
Include this dynamic stylesheet in the `minimal` layout, passing the
organisation's `theme_color`.
Add generated CSS files to `.gitignore`.

This enables organisations to customize the look and feel of embedded event
listings to align with their brand identity.

Also update the event block partial to display the `carousel_text`
directly in an alert box, and update its description.

This makes the `carousel_text` more prominent in event listings.